requirements.txt
# Core dependencies
yt-dlp>=2023.12.30
webvtt-py>=0.4.6
python-docx>=1.1.0
WeasyPrint>=62.0

# Azure services
azure-core>=1.29.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.88.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0 